chore: refactor directory
