+++ tools/ice-scripts/lib/commands/build.js
@@ -7,17 +7,17 @@ const webpack = require('webpack');
 const webpackMerge = require('webpack-merge');
 const { collectDetail } = require('@alifd/fusion-collector');
 
-const paths = require('./config/paths');
-const getBuildConfig = require('./config/getBuildConfig');
-const iceScriptsPkgData = require('../package.json');
-const getWebpackConfigProd = require('./config/webpack.config.prod');
-const goldlog = require('./utils/goldlog');
-const log = require('./utils/log');
-const cliInstance = require('./utils/cliInstance');
-const validationSassAvailable = require('./utils/validationSassAvailable');
-const pkgData = require('./config/packageJson');
-const componentBuild = require('./component/build');
-const checkDepsInstalled = require('./utils/checkDepsInstalled');
+const paths = require('../config/paths');
+const getBuildConfig = require('../config/getBuildConfig');
+const iceScriptsPkgData = require('../../package.json');
+const getWebpackConfigProd = require('../config/webpack.config.prod');
+const goldlog = require('../utils/goldlog');
+const log = require('../utils/log');
+const cliInstance = require('../utils/cliInstance');
+const validationSassAvailable = require('../utils/validationSassAvailable');
+const pkgData = require('../config/packageJson');
+const componentBuild = require('../component/build');
+const checkDepsInstalled = require('../utils/checkDepsInstalled');
 
 /**
  * 构建项目：cli 调用；云构建方法调用
