+++ tools/ice-scripts/lib/config/getRules.js
@@ -18,7 +18,7 @@ const CSS_HOT_LOADER = require.resolve('css-hot-loader');
 const URL_LOADER = require.resolve('url-loader');
 const ICE_SKIN_LOADER = require.resolve('ice-skin-loader');
 const HANDLEBARS_LOADER = require.resolve('handlebars-loader');
-const DEMO_LOADER = require.resolve('../loaders/componentDemoLoader');
+const DEMO_LOADER = require.resolve('../plugins/loaders/componentDemoLoader');
 
 const URL_LOADER_LIMIT = 8192;
