+++ tools/ice-scripts/lib/plugins/loaders/componentDemoLoader.js
