+++ tools/ice-scripts/lib/utils/addBlock.js
