+++ tools/ice-scripts/lib/utils/devMiddleware.js
