+++ tools/ice-scripts/lib/utils/initMaterial.js
