+++ tools/ice-scripts/lib/utils/initProject.js
